modif solution

+++ testMaxValues.cs
@@ -6,7 +6,7 @@ using System;
             int x = int.MaxValue;
             long y = long.MaxValue;
             short z = short.MaxValue;
-            long additionResult = x + y + z;
+            ulong additionResult = (ulong)x + (ulong)y + (ulong)z;
             
             System.Console.WriteLine(additionResult);
         }
